state management modified and moved to Main.jsx. Main.jsx name updated to App.jsx
