+++ package.json
@@ -16,12 +16,12 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "jsoneditor": "^10.0.0",
-    "jsoneditor-react": "^3.1.2",
     "lucide-react": "^0.316.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.21.3",
+    "react18-json-view": "^0.2.8-canary.0",
     "sass": "^1.69.7",
     "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
