+++ src/Components/Sidebar/index.jsx
@@ -1,7 +1,6 @@
 import { Link } from "react-router-dom";
 import { FaHome, FaUser, FaEdit, FaCalendarAlt } from "react-icons/fa";
 import { IoMdSettings } from "react-icons/io";
-import { SiElement } from "react-icons/si";
 import './Sidebar.scss';
 
 const Sidebar = () => {
@@ -12,10 +11,6 @@ const Sidebar = () => {
         <Link to="/" className="sidebar__link" data-testid="sidebar-link-home">
           <FaHome size={25} />
           <p className="sidebar__link-title">Home</p>
-        </Link>
-        <Link to="/app" className="sidebar__link" data-testid="sidebar-link-app">
-          <SiElement size={25} />
-          <p className="sidebar__link-title">App</p>
         </Link>
         <Link to="/" className="sidebar__link" data-testid="sidebar-link-user">
           <FaUser size={25} />
