+++ src/Constants/index.js
@@ -4,10 +4,6 @@ export const navLinks = [
         route: '/',
     },
     {
-        label: 'App',
-        route: '/app',
-    },
-    {
         label: 'History',
         route: '/history',
     },
