--- src/Pages/App/App.scss
@@ -1,12 +1,0 @@
-.app {
-    .app__info {
-        display: flex;
-        align-items: center;
-        gap: 20px;
-        margin-bottom: 20px;
-
-        h1 {
-            font-size: 2em;
-        }
-    }
-}
