+++ src/Pages/Home/index.jsx
@@ -1,17 +1,24 @@
-import './Home.scss';
+import Form from "@/Components/Form";
+import "./Home.scss";
+import Results from "@/Components/Results";
 
-const Home = () => {
+const Home = ({
+  data,
+  requestParams,
+  setRequestParams,
+  isLoading,
+}) => {
   return (
     <div className="home">
-      <div className="box box1">Box 1</div>
-      <div className="box box2">Box 2</div>
-      <div className="box box3">Box 3</div>
-      <div className="box box4">Box 4</div>
-      <div className="box box5">Box 5</div>
-      <div className="box box6">Box 6</div>
-      <div className="box box7">Box 7</div>
-      <div className="box box8">Box 8</div>
-      <div className="box box9">Box 9</div>
+      <div className="home__container">
+        <h1>RESTy</h1>
+        <Form handleApiCall={setRequestParams} />
+      </div>
+      <Results
+        data={data}
+        requestParams={requestParams}
+        isLoading={isLoading}
+      />
     </div>
   );
 };
