+++ src/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import Main from './Main.jsx'
+import App from './App.jsx'
 import './globals.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <Main />
+    <App />
   </React.StrictMode>,
 )
